requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
google-genai>=1.0.0
google-cloud-firestore==2.16.0
Pillow==10.3.0
python-multipart==0.0.9
python-dotenv==1.0.0
